chore: move all options related classes to a new Options module
